refactor!: rename DynamicOutcomeNode to OutcomeNode
